requirements.txt
certifi==2021.10.8
click==8.0.3
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.5
imageio==2.13.5
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.5
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyWavelets==1.2.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.7.0
